+++ pie_extended/models/__init__.py
@@ -2,5 +2,6 @@ modules = [
     "lasla",
     "fro",
     "fr",
-    "freem"
+    "freem",
+    "grc"
 ]